Final Commit
First of all, to access the application, you have to log in and if you do not have an account you have to create it in "register".
after your connection, a list appears containing all the close stores and you have the right to like them or not (dislike them).
Preferred stores will appear in the list of preferred stores and for those you have disliked they will be hidden for two hours.